Improve build 2 (#4)
now `prerelease` and `relaseDraft` in github actions are set as false, and we do not run build in master branch. 
to make it run we should manually upload artifacts using https://github.com/actions/upload-artifact as suggested by Tauri docs but it is too much effort. 

Also fixed name of the app for release and tag is now `noa`.